+++ BACKEND/server/db/database.json
@@ -1,9 +1,9 @@
 {
     "development": {
-      "username": "devdb",
-      "password": "R00t4321#?!",
-      "database": "sagar_test",
-      "host": "172.20.22.20",
+      "username": "root",
+      "password": "P@ssw0rd",
+      "database": "db_balance_humanity",
+      "host": "127.0.0.1",
       "dialect": "mysql"
     },
     "test": {
