+++ BACKEND/server/helpers/minio_helper.js
@@ -1,9 +1,13 @@
 "use strict";
 
-(minioHelper => {
+((minioHelper) => {
     const Minio = require("minio");
     const Promise = require("bluebird");
     const axios = require("axios").default;
+    const fs = require('fs');
+    const { v4: uuidv4 } = require('uuid');
+
+    const httpStatus = require('http-status');
     let minioClient;
     minioHelper.getInstance = () => {
         return minioClient;
@@ -278,7 +282,81 @@
             }
         });
     };
+
+    minioHelper.checkMimeType = (significantBit, fileExtension)=>{
+        let detectedMimeType;
+        console.log('ext: ', fileExtension)
+        switch(significantBit){
+            case "89504e47":
+                detectedMimeType = 'image/png';
+                break;
+            case "ffd8ffe0":
+            case "ffd8ffe1":
+            case "ffd8ffe2":
+            case "ffd8ffe3":
+            case "ffd8ffe8":
+                detectedMimeType = fileExtension =='image/jpeg'?'image/jpeg':'image/jpg';
+                break;
+            default:
+                detectedMimeType = 'uknown';
+                break;
+        }
+        if(detectedMimeType == fileExtension){
+            return {invalid: false}
+        }
+        if(detectedMimeType != fileExtension){
+            return {invalid: true, message:'Only jpeg/jpg and png formats are allowed.'}
+        }
+      }
+
+    minioHelper.uploadImage = async( uploadBucket,file)=>{
+        let response = {status: httpStatus.INTERNAL_SERVER_ERROR, message: `Image upload failed.`}
+        return new Promise((resolve, reject)=>{
+          let fileStream = fs.readFileSync(file.path);
+          let base64 = fileStream.toString('hex');
+          let fileExtension = file.originalname.split('.')[1].toLowerCase();
+          let mimeType = `image/${fileExtension}`;
+          let fileName = `${uuidv4()}.${fileExtension}`
+      
+          let significantBit = base64.slice(0,8)
+          let checkedMimeType = minioHelper.checkMimeType(significantBit, mimeType)
+      
+          if(checkedMimeType && checkedMimeType.invalid){
+            response.status = httpStatus.BAD_REQUEST;
+            response.message = 'Only jpeg/jpg and png formats are allowed.';
+            resolve(response);
+          }
+      
+          let metaData = {
+            'Content-Type': mimeType
+          }
+      
+          fs.stat(file.path, function(statError, stats){
+              if(statError){
+                resolve(response);
+              }
+        
+              minioClient.putObject(uploadBucket, fileName, fileStream, stats.size, metaData, function(uploadingError, etag){
+                if(uploadingError){
+                  resolve(response);
+                }
+        
+               let data = {
+                  url:`${process.env.MINIO_BASE_URL}${uploadBucket}/${fileName}`,
+                  info: {...etag, fileName}
+                }
+                response = {status:200, data}
+                resolve(response);
+              } )
+          })
+        })
+      }
+
 })(module.exports);
+
+
+
+
 // hint for image upload in minio
 // const file = fs.readFileSync('/home/manoj/bfi-tech/backend/common/mcs-common-service-node/server/assets/Global_IME_logo.png')
 //   let metaData = {
